auth-service/requirements.txt
Django>=5.0
djangorestframework
requests
PyJWT
mysqlclient
